requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sentence-transformers
faiss-cpu
openai
python-dotenv
lxml
